+++ src/main/kotlin/me/fzzyhmstrs/lootables/command/LootableTableArgumentType.kt
@@ -16,7 +16,7 @@ import com.mojang.brigadier.context.CommandContext
 import com.mojang.brigadier.exceptions.DynamicCommandExceptionType
 import com.mojang.brigadier.suggestion.Suggestions
 import com.mojang.brigadier.suggestion.SuggestionsBuilder
-import me.fzzyhmstrs.lootables.loot.LootablesData
+import me.fzzyhmstrs.lootables.data.LootablesData
 import net.minecraft.command.CommandSource
 import net.minecraft.server.command.ServerCommandSource
 import net.minecraft.text.Text