+++ src/main/kotlin/me/fzzyhmstrs/lootables/lootables.kt
@@ -15,7 +15,7 @@ package me.fzzyhmstrs.lootables
 import me.fzzyhmstrs.lootables.command.LootablesCommands
 import me.fzzyhmstrs.lootables.config.LootablesConfig
 import me.fzzyhmstrs.lootables.loot.LootablePoolEntryTypes
-import me.fzzyhmstrs.lootables.loot.LootablesData
+import me.fzzyhmstrs.lootables.data.LootablesData
 import me.fzzyhmstrs.lootables.network.LootablesNetworking
 import me.fzzyhmstrs.lootables.registry.ComponentRegistry
 import me.fzzyhmstrs.lootables.registry.LootFunctionRegistry