+++ changelog.md
@@ -1,8 +1,11 @@
 ### Additions
-* Initial release.
+* None.
 
 ### Changes
-* None.
+* Lootable tables now reload (async) on /reload. Previously they only loaded on server start.
+* Sync data now supports player-specific client data
 
 ### Fixes
-* None.
+* Fix loot table entries not loading properly.
+* Pool loot entries properly display their fallback description.
+* Fix potential sync issue on quit and rejoin thanks to stale registry references.